requirements.txt
Flask==1.1.2
Flask-Cors
Flask-JWT-Simple==0.0.3
Flask-RESTful==0.3.8
future==0.17.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
json5==0.9.5
jsonpickle==1.4.1
jsonschema==2.6.0
logger==1.4
oauthlib==3.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
PyJWT==1.7.1
pylint
python-crfsuite==0.9.7
python-engineio==3.13.2
python-jsonrpc-server
python-socketio==3.1.2
python-telegram-bot==11.1.0
PyYAML==5.3.1
requests
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rocketchat-API==0.6.36
sklearn_crfsuite==0.3.6
simplejson==3.17.2
Sphinx
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport
tornado==5.1.1
ujson==1.35
unicodecsv==0.14.1
websocket-client==0.54.0
Werkzeug==1.0.1
